refactor(phase-3): relocate domain components into features and components/layout
