+++ app/(main)/[locale]/_components/homepage-client.tsx
