+++ features/events/components/onboarding.tsx
